Tussencommit Abonnement View aanpassen
